requirements.txt
# Core
numpy>=1.23
tqdm>=4.66
PyYAML>=6.0
scikit-learn>=1.3
pandas>=2.0
rich>=13.0

# Deep Learning
torch>=2.1
torchvision>=0.16
timm>=0.9.12

# Dataset
medmnist>=3.0.0

# Logging
tensorboard>=2.14
